refactor(tests): complete migration to cli_fixtures.strip_ansi_codes
- Migrate all 16 remaining test files from tests.utils to tests.cli_fixtures import
- Update documentation in tests/CLAUDE.md to reflect new import location
- Add deprecation warning to tests.utils.strip_ansi_codes wrapper
- Internal CLITestHelper uses direct import to avoid deprecation warnings

This completes the full migration recommended in PR #356 review.
All tests pass with the new import structure.
