+++ tests/integration/test_cli_index.py
@@ -9,7 +9,7 @@ from typer.testing import CliRunner
 
 from scriptrag.cli.main import app
 from scriptrag.config import ScriptRAGSettings, set_settings
-from tests.utils import strip_ansi_codes
+from tests.cli_fixtures import strip_ansi_codes
 
 runner = CliRunner()
