+++ tests/integration/test_cli_list.py
@@ -7,7 +7,7 @@ import pytest
 from typer.testing import CliRunner
 
 from scriptrag.cli import app
-from tests.utils import strip_ansi_codes
+from tests.cli_fixtures import strip_ansi_codes
 
 # Path to fixture files
 FIXTURES_DIR = Path(__file__).parent.parent / "fixtures" / "fountain"
