+++ tests/integration/test_context_query.py
@@ -13,7 +13,7 @@ from typer.testing import CliRunner
 
 from scriptrag.cli.main import app
 from scriptrag.config import set_settings
-from tests.utils import strip_ansi_codes
+from tests.cli_fixtures import strip_ansi_codes
 
 runner = CliRunner()
