+++ tests/unit/test_cli_commands_scene.py
@@ -16,7 +16,7 @@ from scriptrag.api.scene_models import (
 )
 from scriptrag.cli.main import app
 from scriptrag.parser import Scene
-from tests.utils import strip_ansi_codes
+from tests.cli_fixtures import strip_ansi_codes
 
 
 class TestSceneCommandsConfigOption:
